Support 'ActualText' property
When a PDF uses embedded fonts that do NOT use a standard named encoding
or the charcters in the fonts are NOT identified by a standard named or
CIDs in a well-known collectin, the PDF may contain optional 'ToUnicode'
entries in the font directory and 'ActualText' entries for structured
elements or marked-content sequences in order to specify text for text
extraction. (See at 9.10.1 in the specification of PDF 1.7)

In the Adobe reader, when both 'ToUnicode' and 'ActualText' are
spedified, it will use the information in 'ActualText' entries. This
commit implements the same behavior in PDF.js.